requirements.txt
flask==3.0.2
requests==2.31.0
pandas==2.2.1
apify-client==1.6.1
python-dotenv==1.0.1 